Created the allocations view and budget manipulation view and the services related to it

+++ schema/allocation.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 
 class AllocationCreate(BaseModel):
     category_id: int
-    allocated_amount: Decimal
+    allocated_amount: float
 
 
 class AllocationOut(AllocationCreate):
